% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tune_optuna.R
\name{as_kernels.random_machines_tune}
\alias{as_kernels.random_machines_tune}
\title{Extract Best Kernels from Grid Search Results}
\usage{
\method{as_kernels}{random_machines_tune}(tune_results, kernel_mix)
}
\arguments{
\item{tune_results}{The object returned by \code{tune_random_machines}.}

\item{kernel_mix}{Original kernel configuration.}
}
\description{
Helper function to merge the original kernel configuration with the 
best hyperparameters found by standard Grid Search (\code{tune_random_machines}).
}
