% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tune.R, R/tune_optuna.R
\name{as_kernels}
\alias{as_kernels}
\title{Prepare Tuned Kernels for Random Machines}
\usage{
as_kernels(tune_results, kernel_mix)

as_kernels(tune_results, kernel_mix)
}
\arguments{
\item{tune_results}{Tuning result object.}

\item{kernel_mix}{Original kernel configuration.}
}
\description{
Defines the generic function \code{as_kernels} to extract optimized kernel
configurations from tuning results.
}
