% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tune.R
\name{$.fastsvm_custom_kernel}
\alias{$.fastsvm_custom_kernel}
\title{Accessor for Custom Kernel Parameters}
\usage{
\method{$}{fastsvm_custom_kernel}(x, name)
}
\arguments{
\item{x}{A custom kernel object.}

\item{name}{The name of the parameter to access.}
}
\description{
Accessor for Custom Kernel Parameters
}
