% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tune.R
\name{create_kernel_variants}
\alias{create_kernel_variants}
\title{Internal helper to create variants of a custom kernel function}
\usage{
create_kernel_variants(kernel_factory, ...)
}
\description{
Internal helper to create variants of a custom kernel function
}
\keyword{internal}
