% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_machines.R
\name{.compute_lambdas_from_cindex}
\alias{.compute_lambdas_from_cindex}
\title{Compute normalized weights from C-indices}
\usage{
.compute_lambdas_from_cindex(cindex_vec, beta = 1)
}
\description{
Compute normalized weights from C-indices
}
\keyword{internal}
