% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tune_optuna.R
\name{opt_cat}
\alias{opt_cat}
\title{Define Categorical Parameter for Optuna Search Space}
\usage{
opt_cat(choices)
}
\arguments{
\item{choices}{A vector of values (numeric, character, or logical) to choose from.}
}
\value{
An object of class \code{"opt_param_def"}.
}
\description{
Helper function to define a categorical choice set for Bayesian optimization.
}
\examples{
# Choose between different optimizers
space_opt <- opt_cat(c("avltree", "rbtree"))
}
