% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bagging.R
\name{predict.fastsvm_bag}
\alias{predict.fastsvm_bag}
\title{Predict method (Optional - for new data)
Uses serialized models.}
\usage{
\method{predict}{fastsvm_bag}(object, newdata, ...)
}
\description{
Predict method (Optional - for new data)
Uses serialized models.
}
