% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastsvm.R
\name{print.fastsvm}
\alias{print.fastsvm}
\title{Print method for fastsvm objects}
\usage{
\method{print}{fastsvm}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"fastsvm"}.}

\item{...}{Not used.}
}
\description{
This method prints a compact summary of the fitted
\code{"fastsvm"} model, including the number of observations,
number of covariates, kernel type, and key hyperparameters.
}
