% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tune.R
\name{print.fastsvm_multi_grid}
\alias{print.fastsvm_multi_grid}
\title{Print method for multi-kernel results}
\usage{
\method{print}{fastsvm_multi_grid}(x, ...)
}
\description{
Print method for multi-kernel results
}
