% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fast_kernel_surv_svm.R
\name{print.summary.fastsvm}
\alias{print.summary.fastsvm}
\title{Print method for summary.fastsvm objects}
\usage{
\method{print}{summary.fastsvm}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"summary.fastsvm"}.}

\item{...}{Not used.}
}
\description{
Print method for summary.fastsvm objects
}
