% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_machines.R
\name{score.random_machines}
\alias{score.random_machines}
\title{Score method for Random Machines}
\usage{
\method{score}{random_machines}(object, data, ...)
}
\arguments{
\item{object}{A fitted \code{random_machines} model.}

\item{data}{Validation data containing time and event columns.}

\item{...}{Additional arguments.}
}
\description{
Score method for Random Machines
}
