% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_machines.R
\name{score}
\alias{score}
\title{Generic function for computing concordance index (score)}
\usage{
score(object, data, ...)
}
\arguments{
\item{object}{A fitted model object.}

\item{data}{A data frame containing the validation data.}

\item{...}{Additional arguments.}
}
\description{
Generic function for computing concordance index (score)
}
