% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bagging.R
\name{score_fastsvm_bag}
\alias{score_fastsvm_bag}
\title{Score method for Bagging Result}
\usage{
score_fastsvm_bag(object, data)
}
\arguments{
\item{object}{An object of class \code{"fastsvm_bag"}.}

\item{data}{Validation data (must contain time/event cols).}
}
\description{
Computes the concordance index for the aggregated predictions.
}
