% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastsvm.R
\name{summary.fastsvm}
\alias{summary.fastsvm}
\title{Summary method for fastsvm objects}
\usage{
\method{summary}{fastsvm}(object, ...)
}
\arguments{
\item{object}{An object of class \code{"fastsvm"}.}

\item{...}{Not used.}
}
\value{
An object of class \code{"summary.fastsvm"}.
}
\description{
Extracts key information from a fitted \code{"fastsvm"} model, including
the number of observations, number of covariates, kernel specification,
hyperparameters, and a summary of the sample-wise coefficients
\code{coef_} returned by the Python model.
}
