% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tune.R
\name{as_kernels}
\alias{as_kernels}
\title{Prepare Tuned Kernels for Random Machines}
\usage{
as_kernels(tune_results, kernel_mix)
}
\arguments{
\item{tune_results}{The results from \code{tune_random_machines}.}

\item{kernel_mix}{The original base configuration list.}
}
\description{
Prepare Tuned Kernels for Random Machines
}
