% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_machines.R
\name{predict.random_machines}
\alias{predict.random_machines}
\title{Predict method for Random Machines}
\usage{
\method{predict}{random_machines}(object, newdata, ...)
}
\arguments{
\item{object}{An object of class \code{"random_machines"}.}

\item{newdata}{Data frame with new observations.}

\item{...}{Additional arguments.}
}
\description{
Predict method for Random Machines
}
