% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tune.R
\name{print.fastsvm_custom_kernel}
\alias{print.fastsvm_custom_kernel}
\title{Print method for custom kernels}
\usage{
\method{print}{fastsvm_custom_kernel}(x, ...)
}
\arguments{
\item{x}{A custom kernel object.}

\item{...}{Additional arguments.}
}
\description{
Print method for custom kernels
}
