% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_machines.R
\name{score.random_machines}
\alias{score.random_machines}
\title{Score method for Random Machines}
\usage{
\method{score}{random_machines}(object, data, ...)
}
\arguments{
\item{object}{An object of class \code{"random_machines"}.}

\item{data}{Validation data (must contain time/event cols).}

\item{...}{Not used.}
}
\description{
Computes the concordance index for the aggregated predictions.
}
