% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tune.R
\name{tune_fastsvm}
\alias{tune_fastsvm}
\title{Single Grid Search for FastKernelSurvivalSVM (Internal)}
\usage{
tune_fastsvm(
  data,
  time_col = "t",
  delta_col = "delta",
  param_grid,
  cv = 5L,
  n_jobs = 1L,
  verbose = 0L,
  refit = TRUE,
  ...
)
}
\arguments{
\item{data}{A \code{data.frame} with training data.}

\item{time_col}{Name of the time column.}

\item{delta_col}{Name of the event column.}

\item{param_grid}{A list (or list of lists) of parameters to tune.}

\item{cv}{Integer. Number of folds.}

\item{n_jobs}{Integer. Use 1 if using custom kernels.}

\item{verbose}{Integer.}

\item{...}{Additional fixed parameters (e.g., \code{max_iter}).}
}
\description{
Wrapper for \code{sklearn.model_selection.GridSearchCV}.
}
